requirements.txt
google-cloud-vision>=3.4.4
google-generativeai>=0.7.0
pillow-heif>=0.13.0
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
tqdm>=4.65.0
imagehash>=4.3.1
scikit-image>=0.21.0
numpy>=1.24.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
rapidfuzz>=3.2.0
pydantic>=2.0.0
requests>=2.31.0
unicodedata2>=15.0.0